+++ test/rnn_vanilla_common.hpp
@@ -234,7 +234,7 @@ void RNNFwdTrainCPUVerify(miopen::Handle& handle,
             if(use_dropout)
             {
                 auto dropout_states_tmp = dropout_states_host;
-                size_t drop_out_offset  = (li - 1) * batch_n * hy_h * bi;
+                size_t drop_out_offset  = (li - 1ULL) * batch_n * hy_h * bi;
 
                 DropoutForwardVerify<T>(handle,
                                         dropoutDesc,
@@ -2464,7 +2464,7 @@ struct rnn_basic_vanilla_driver : test_driver
         // Create input tensor
         // If we are in skip mode, take the real input size to be the vector length.
         auto inVecReal    = (inputMode != 0) ? hiddenSize : inVecLen;
-        std::size_t in_sz = inVecReal * batch_n;
+        std::size_t in_sz = static_cast<std::size_t>(inVecReal) * batch_n;
         std::vector<T> input(in_sz);
         srand(0);
         for(std::size_t i = 0; i < in_sz; i++)
@@ -2472,7 +2472,7 @@ struct rnn_basic_vanilla_driver : test_driver
             input[i] = /*(((GET_RAND()%2)==1)?-1:1)**/ 0.001 * float(GET_RAND() % 100);
         }
 
-        std::size_t hx_sz = ((dirMode != 0) ? 2 : 1) * hiddenSize * batchSize * numLayers;
+        std::size_t hx_sz = ((dirMode != 0) ? 2ULL : 1ULL) * hiddenSize * batchSize * numLayers;
         std::vector<T> hx;
         if(!nohx)
             hx.resize(hx_sz);
